requirements.txt
fastapi
uvicorn
python-telegram-bot
sqlalchemy
psycopg2-binary
requests
apscheduler
jinja2
python-dotenv
auth
